feat: reorganise analysis functions
This reorganises them and makes the individual reporters act like
"plugins" so we can add more in future.
